tools/release/pyproject.toml
[build-system]
requires = ["flit_core>=3,<4"]
build-backend = "flit_core.buildapi"

[project]
authors = [{ name = "Planetarium", email = "engineering@planetariumhq.com" }]
name = "toolbelt"
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "py7zr ~=0.20.0",
    "structlog >=16.1,<22.1",
    "boto3 ~=1.24",
    "PyYAML ==6.0",
    "requests ==2.26.0",
    "typer[all] ~=0.6.1",
    "python-dotenv ~=0.19",
    "ruamel.yaml ~=0.17.21"
]
dynamic = ["version", "description"]

[project.optional-dependencies]
test = [
    "pytest >=7.0,<7.2",
    "pytest-mock >=3.6.0,<3.10",
    "pytest-cov >=2.10.0,<3.0.0",
    "requests-mock~=1.10.0",
]
type = ["mypy ==0.971", "types-requests", "types-PyYAML"]
dev = [
    "pre-commit >=2.17.0,<3.0.0",
    "colorama >=0.4.3,<0.5.0",
    "black ~=22.6.0",
    "isort >=5.0.6,<6.0.0",
]

[tool.flit.sdist]
include = ["toolbelt/*"]
exclude = [".env.sample", ".vscode/*", "tests/*", "scripts/*"]

[tool.flit.module]
name = "toolbelt"

[tool.black]
line-length = 88

[tool.pytest.ini_options]
addopts = "--cov=toolbelt"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
skip_gitignore = true
